requirements.txt
prisma 
fastapi
uvicorn
python-dotenv
bcrypt
PyJWT
pydantic
pydantic[email]
python-multipart